+++ src/backends/sentry_backend_inproc.c
@@ -1,3 +1,4 @@
+#include "sentry_attachment.h"
 #include "sentry_boot.h"
 
 #include "sentry_alloc.h"
@@ -601,8 +602,8 @@ handle_ucontext(const sentry_ucontext_t *uctx)
                 sentry_path_t *screenshot_path
                     = sentry__screenshot_get_path(options);
                 if (sentry__screenshot_capture(screenshot_path)) {
-                    sentry__envelope_add_attachment(
-                        envelope, screenshot_path, ATTACHMENT, NULL);
+                    sentry__envelope_add_attachment(envelope,
+                        sentry__attachment_from_path(screenshot_path));
                 }
                 sentry__path_free(screenshot_path);
             }