+++ src/sentry_core.c
@@ -1428,6 +1428,21 @@ sentry_capture_minidump_n(const char *path, size_t path_len)
     return sentry_uuid_nil();
 }
 
+static sentry_attachment_t *
+add_attachment(sentry_attachment_t *attachment)
+{
+    SENTRY_WITH_OPTIONS (options) {
+        if (options->backend && options->backend->add_attachment_func) {
+            options->backend->add_attachment_func(options->backend, attachment);
+        }
+    }
+    SENTRY_WITH_SCOPE_MUT (scope) {
+        return sentry__attachments_add(
+            &scope->attachments, attachment, ATTACHMENT, NULL);
+    }
+    return NULL;
+}
+
 sentry_attachment_t *
 sentry_attach_file(const char *path)
 {
@@ -1437,19 +1452,23 @@ sentry_attach_file(const char *path)
 sentry_attachment_t *
 sentry_attach_file_n(const char *path, size_t path_len)
 {
-    sentry_path_t *attachment_path = sentry__path_from_str_n(path, path_len);
-    SENTRY_WITH_OPTIONS (options) {
-        if (options->backend && options->backend->add_attachment_func) {
-            options->backend->add_attachment_func(
-                options->backend, attachment_path);
-        }
-    }
-    sentry_attachment_t *attachment = NULL;
-    SENTRY_WITH_SCOPE_MUT (scope) {
-        attachment = sentry__attachments_add(
-            &scope->attachments, attachment_path, ATTACHMENT, NULL);
-    }
-    return attachment;
+    return add_attachment(
+        sentry__attachment_from_path(sentry__path_from_str_n(path, path_len)));
+}
+
+sentry_attachment_t *
+sentry_attach_bytes(const char *buf, size_t buf_len, const char *filename)
+{
+    return sentry_attach_bytes_n(
+        buf, buf_len, filename, sentry__guarded_strlen(filename));
+}
+
+sentry_attachment_t *
+sentry_attach_bytes_n(
+    const char *buf, size_t buf_len, const char *filename, size_t filename_len)
+{
+    return add_attachment(sentry__attachment_from_buffer(
+        buf, buf_len, sentry__path_from_str_n(filename, filename_len)));
 }
 
 void
@@ -1458,7 +1477,7 @@ sentry_remove_attachment(sentry_attachment_t *attachment)
     SENTRY_WITH_OPTIONS (options) {
         if (options->backend && options->backend->remove_attachment_func) {
             options->backend->remove_attachment_func(
-                options->backend, attachment->path);
+                options->backend, attachment);
         }
     }
     SENTRY_WITH_SCOPE_MUT (scope) {
@@ -1477,18 +1496,22 @@ sentry_attach_filew(const wchar_t *path)
 sentry_attachment_t *
 sentry_attach_filew_n(const wchar_t *path, size_t path_len)
 {
-    sentry_path_t *attachment_path = sentry__path_from_wstr_n(path, path_len);
-    SENTRY_WITH_OPTIONS (options) {
-        if (options->backend && options->backend->add_attachment_func) {
-            options->backend->add_attachment_func(
-                options->backend, attachment_path);
-        }
-    }
-    sentry_attachment_t *attachment = NULL;
-    SENTRY_WITH_SCOPE_MUT (scope) {
-        attachment = sentry__attachments_add(
-            &scope->attachments, attachment_path, ATTACHMENT, NULL);
-    }
-    return attachment;
+    return add_attachment(sentry__attachment_from_path(
+        sentry__path_from_wstr_n(path, path_len)));
+}
+
+sentry_attachment_t *
+sentry_attach_bytesw(const char *buf, size_t buf_len, const wchar_t *filename)
+{
+    size_t filename_len = filename ? wcslen(filename) : 0;
+    return sentry_attach_bytesw_n(buf, buf_len, filename, filename_len);
+}
+
+sentry_attachment_t *
+sentry_attach_bytesw_n(const char *buf, size_t buf_len, const wchar_t *filename,
+    size_t filename_len)
+{
+    return add_attachment(sentry__attachment_from_buffer(
+        buf, buf_len, sentry__path_from_wstr_n(filename, filename_len)));
 }
 #endif